+++ src/android/src/de/akaflieg_freiburg/enroute/MobileAdaptor.java
@@ -54,7 +54,6 @@ public class MobileAdaptor extends de.akaflieg_freiburg.enroute.ShareActivity {
 
   public static native void onLanguageChanged();
   public static native void onWifiConnected();
-  public static native void onWindowSizeChanged();
 
   private static MobileAdaptor m_instance;
   private static Vibrator m_vibrator;
@@ -111,22 +110,6 @@ public class MobileAdaptor extends de.akaflieg_freiburg.enroute.ShareActivity {
       IntentFilter filter = new IntentFilter(Intent.ACTION_LOCALE_CHANGED);
       registerReceiver(m_localeChangedReceiver, filter);
     }
-
-    // Be informed when the window size changes, and call the C++ method
-    // onWindowSizeChanged() whenever it changes. The window size changes when
-    // the user starts/end the split view mode, or when the user drags the
-    // slider in order to adjust the relative size of the two windows shown.
-    View rootView = getWindow().getDecorView().getRootView();
-    rootView.addOnLayoutChangeListener(new View.OnLayoutChangeListener() 
-      {
-        @Override
-        public void onLayoutChange(View view, int left, int top, int right, int bottom, int oldLeft,
-            int oldTop, int oldRight, int oldBottom) 
-        {
-          onWindowSizeChanged();
-        }
-      }
-    );
   }
 
   @Override
@@ -167,62 +150,6 @@ public class MobileAdaptor extends de.akaflieg_freiburg.enroute.ShareActivity {
   public static String deviceName() {
     return android.os.Build.MANUFACTURER + " " + android.os.Build.PRODUCT + " ("
         + android.os.Build.MODEL + ")";
-  }
-
-  // Returns the height of the screen, taking the Android split view
-  // into account
-  public static double windowHeight() {
-    return m_instance.getWindow().getDecorView().getRootView().getHeight();
-  }
-
-  // Returns the width of the screen, taking the Android split view
-  // into account
-  public static double windowWidth() {
-    return m_instance.getWindow().getDecorView().getRootView().getWidth();
-  }
-
-  // Returns the bottom inset required to avoid system bars and display cutouts
-  public static double safeInsetBottom() {
-    if (Build.VERSION.SDK_INT >= 30) {
-      return m_instance.getWindow().getDecorView().getRootWindowInsets()
-          .getInsets(WindowInsets.Type.systemBars() | WindowInsets.Type.ime()
-              | WindowInsets.Type.displayCutout()).bottom;
-    }
-
-    return m_instance.getWindow().getDecorView().getRootWindowInsets().getSystemWindowInsetBottom();
-  }
-
-  // Returns the left inset required to avoid system bars and display cutouts
-  public static double safeInsetLeft() {
-    if (Build.VERSION.SDK_INT >= 30) {
-      return m_instance.getWindow().getDecorView().getRootWindowInsets()
-          .getInsets(WindowInsets.Type.systemBars() | WindowInsets.Type.ime()
-              | WindowInsets.Type.displayCutout()).left;
-    }
-
-    return m_instance.getWindow().getDecorView().getRootWindowInsets().getSystemWindowInsetLeft();
-  }
-
-  // Returns the right inset required to avoid system bars and display cutouts
-  public static double safeInsetRight() {
-    if (Build.VERSION.SDK_INT >= 30) {
-      return m_instance.getWindow().getDecorView().getRootWindowInsets()
-          .getInsets(WindowInsets.Type.systemBars() | WindowInsets.Type.ime()
-              | WindowInsets.Type.displayCutout()).right;
-    }
-
-    return m_instance.getWindow().getDecorView().getRootWindowInsets().getSystemWindowInsetRight();
-  }
-
-  // Returns the top inset required to avoid system bars and display cutouts
-  public static double safeInsetTop() {
-    if (Build.VERSION.SDK_INT >= 30) {
-      return m_instance.getWindow().getDecorView().getRootWindowInsets()
-          .getInsets(WindowInsets.Type.systemBars() | WindowInsets.Type.ime()
-              | WindowInsets.Type.displayCutout()).top;
-    }
-
-    return m_instance.getWindow().getDecorView().getRootWindowInsets().getSystemWindowInsetTop();
   }
 
   /*